Roman Numeral Conversion Program
Prompt: Write a program that converts a number entered in Roman Numerals to a decimal. Your program should consist of a class, say DecimalClass. An object of type DecimalClass should do the following:

a. Store the number as a Roman numeral. 
b. Convert and store the number into decimal form. 
c. Print the number as a Roman numeral or decimal number as requested by the user. 

The decimal values of the Roman Numerals are:

M = 1000
D = 500
C = 100
L = 50
X = 10
V = 5
I = 1